pyproject.toml
[tool.poetry]
name = "TeaRCON"
version = "2.0"
description = "Бот для Telegeram позволяющий выполнять команды на сервере Minecraft по RCON"
authors = ["TEANUS <teanus@mail.ru>", "Taskovich"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
aiogram = "3.4.1"
aiosqlite = "0.20.0"
asyncpg = "0.29.0"
python-dotenv = "1.0.1"
mcrcon = "0.7.0"
pyyaml = "6.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
